♻️ refactor(server): MulticastMdnsResponder → commonMain on the MdnsSocketLayer seam
